ai-agent/requirements.txt
fastapi
uvicorn
langgraph
langchain-openai
langchain-community
tavily-python
pydantic-settings
httpx
python-dotenv